apps/indexer/requirements.txt
# Vector store
qdrant-client>=1.7.0

# Embeddings
openai>=1.10.0
sentence-transformers>=2.3.0

# Document processing
langchain>=0.1.0
langchain-community>=0.0.10

# GitHub
PyGithub>=2.1.1

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
schedule>=1.2.0
